+++ HISTORY.md
@@ -1,8 +1,9 @@
 # History
 
 ## Unreleased
-* Add tests for python 3.15 (alpha)
 * Add support for multiple response variations based on `Vary`
+* Fix overwriting headers after revalidation in `read_only` mode
+* Add tests for python 3.15 (alpha)
 
 ## 1.3.0 (2026-02-02)